requirements.txt
absl-py
tqdm
numpy>=1.18.3
carla==0.9.13
carla-birdeye-view==1.1.1
shapely
aiofiles==22.1.0
protobuf==3.20.1
ray[rllib]==2.0.0
gym==0.21.0
opencv-python==4.2.0.34
mlflow==1.26.0
boto3==1.24.71
psutil